+++ changelog.md
@@ -7,3 +7,10 @@ v0.0
 - added menu file with functions
 - added map file for the different maps of the game
 - added info file for info about characters, locations, and items
+
+v1.0
+- changed the main game file to have allow for user input for movement
+- changed map file to have classes to create the different maps
+- added item file with classes to store info about different type of items
+- added villian file for all the enemies that can damage the player
+- added player file for the characteristics of the player